requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Werkzeug==3.0.0
